+++ openspec/changes/civ7-support-direct-control-modularization/tasks.md
@@ -500,6 +500,13 @@ runtime/direct-control claims.
         operation/proof telemetry, hotseat runtime proof, AI ingestion, CLI
         semantic projection, Effect/oRPC procedure-core work, and Task 2.9.4
         matrix-row acceptance pending.
+  - [x] 4.8.28 Extract diplomacy-response type owner while keeping public
+        facade type re-exports in `index.ts`, preserving diplomacy response
+        input, command payload, result, and postcondition public contracts
+        while leaving narrative closeout, public procedure schemas,
+        operation/proof telemetry, hotseat runtime proof, AI ingestion, CLI
+        semantic projection, Effect/oRPC procedure-core work, and Task 2.9.4
+        matrix-row acceptance pending.
 - [ ] 4.9 Import or explicitly cite oRPC architecture authority before any
       oRPC implementation; current support branch has no tracked
       `.agents/skills/civ7-orpc-control-architecture` or
